Modified HOME_DIR_HACK code section to return error messages to the calling routines, and moved code section to the beginning of dctsearch.
Added the ability to check for shell environment variables
JGOFS_OBJECT and JGOFS_METHOD which can be used to circumvent
the /etc/passwd requirement for a 'jgofs' user.  If these
don't exist, and there isn't a user 'jgofs' defined an error
will be reported to the server, and reported in the httpd logs.