Only check for the ssl certificate and key files if we are a secure connection.
If an application connected to the BES but does not send the
PPT_TESTING_CONNECTION string, then the BES was hanging. Using a timeout for
the initial handshake.

If a client connects to the BES that is not a BES client, and sends a string
other then PPT_TESTING_CONNECTION, then we simply want to close the
connection and return rather than throwing an exception and exiting.
M    ppt/PPTServer.h
M    ppt/PPTClient.cc
M    ppt/SSLConnection.cc
M    ppt/PPTServer.cc
M    ppt/PPTConnection.cc
M    ppt/PPTConnection.h