o Fixed seg fault in isScopeX functions for _scope.empty() case.
o Implemented task 1322.  We can now handle renaming atomic and container attributes if the attribute@orgName is specified.
NOTE: There is a memory leak in AttrTable we are bumping into now (ticket 1326).  I added a ticket 1327 to ncml_handler as well to make sure this is cleaned up.

o New make check test case for changing container and atomic attribute.